if this works

+++ python_variable_annotations/7-to_kv.py
@@ -3,7 +3,7 @@
 from typing import Union, Tuple
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """This is a toople"""
     kv_tuple: Tuple = (k, v*v)
     return kv_tuple
